+++ spec/rails23_spec.rb
@@ -5,7 +5,6 @@ describe "Rails 2.3.x" do
     app = Hatchet::Runner.new('rails23_mri_187').setup!
     app.heroku.put_stack(app.name, "cedar")
     app.deploy do |app, heroku|
-      add_database(app, heroku)
       expect(successful_body(app)).to eq("hello")
     end
   end