+++ spec/rails3_spec.rb
@@ -4,7 +4,6 @@ describe "Rails 3.x" do
   it "should deploy on ruby 1.9.3" do
     Hatchet::Runner.new("rails3_mri_193").deploy do |app, heroku|
       expect(app.output).to include("Asset precompilation completed")
-      add_database(app, heroku)
 
       expect(app.output).to match("WARNING")
       expect(app.output).to match("Add 'rails_12factor' gem to your Gemfile to skip plugin injection")
@@ -19,7 +18,6 @@ describe "Rails 3.x" do
 
   it "should not have warnings when using the rails_12factor gem" do
     Hatchet::Runner.new("rails3_12factor").deploy do |app, heroku|
-      add_database(app, heroku)
       expect(app.output).not_to match("Add 'rails_12factor' gem to your Gemfile to skip plugin injection")
       expect(successful_body(app)).to eq("hello")
     end
@@ -27,7 +25,6 @@ describe "Rails 3.x" do
 
   it "should only display the correct plugin warning" do
     Hatchet::Runner.new("rails3_one_plugin").deploy do |app, heroku|
-      add_database(app, heroku)
       expect(app.output).not_to match("rails_log_stdout")
       expect(app.output).to match("rails3_serve_static_assets")
       expect(app.output).to match("Add 'rails_12factor' gem to your Gemfile to skip plugin injection")
@@ -39,7 +36,6 @@ describe "Rails 3.x" do
     it "should deploy on ruby 1.9.3" do
       Hatchet::Runner.new("railties3_mri_193").deploy do |app, heroku|
         expect(app.output).to include("Asset precompilation completed")
-        add_database(app, heroku)
         expect(app.output).to match("Ruby/Rails")
         expect(successful_body(app)).to eq("hello")
       end