+++ spec/ruby_spec.rb
@@ -3,6 +3,7 @@ require_relative 'spec_helper'
 describe "Ruby apps" do
   describe "default WEB_CONCURRENCY" do
     it "auto scales WEB_CONCURRENCY" do
+      pending("https://github.com/heroku/api/issues/4426")
       app = Hatchet::Runner.new("default_ruby")
       app.setup!
       app.set_config("SENSIBLE_DEFAULTS" => "enabled")
@@ -55,6 +56,24 @@ describe "Ruby apps" do
       it "runs a rake task if the gem exists" do
         Hatchet::Runner.new('mri_200_rake').deploy do |app, heroku|
           expect(app.output).to include("foo")
+        end
+      end
+    end
+  end
+
+  describe "database configuration" do
+    context "no active record" do
+      it "writes a heroku specific database.yml" do
+        Hatchet::Runner.new("default_ruby").deploy do |app, heroku|
+          expect(app.output).to include("Writing config/database.yml to read from DATABASE_URL")
+        end
+      end
+    end
+
+    context "active record 4.1+" do
+      it "doesn't write a heroku specific database.yml" do
+        Hatchet::Runner.new("activerecord41_scaffold").deploy do |app, heroku|
+          expect(app.output).not_to include("Writing config/database.yml to read from DATABASE_URL")
         end
       end
     end